Combine code from image file service and image loader into one class. Reduces code.
