+++ PittsburghCityBridges/Services/ImageFileService.swift
@@ -12,6 +12,7 @@ import os
 class ImageFileService: ObservableObject {
     private let bridgeImagesFolderName = "BridgeImages"
     private var bridgeImagesFolder: Folder?
+    private (set)var cachedImageData: Data?
     private var cachesFolder: Folder?
     
     let logger: Logger = Logger(subsystem: AppLogging.subsystem, category: AppLogging.debugging)
@@ -36,6 +37,33 @@ class ImageFileService: ObservableObject {
                 fatalError("\(#file) \(#function) could not create bridgeImagesFolder")
             }
         }
+    }
+    
+    @MainActor
+    func getImageData(for imageURL: URL?) async -> Data? {
+        guard let imageURL = imageURL else { return nil }
+        var imageData: Data?
+        if let cachedImageData = cachedImageData {
+            return cachedImageData
+        }
+        let imageFileName = imageFileName(imageURL)
+        let imageFileData = getData(for: imageFileName)
+        if let imageFileData = imageFileData {
+            imageData = imageFileData
+            cachedImageData = imageData
+        } else {
+            do {
+                let (data, response) = try await URLSession.shared.data(from: imageURL)
+                logger.debug("\(#file) \(#function) \(response.debugDescription)")
+                // persist, e.g. cache this data for future use
+                imageData = data
+                cachedImageData = data
+                save(data: data, to: imageFileName)
+            } catch {
+                logger.error("\(error.localizedDescription)")
+            }
+        }
+        return imageData
     }
     
     
@@ -76,6 +104,15 @@ class ImageFileService: ObservableObject {
         } catch {
             logger.info("\(#file) \(#function) error \(error.localizedDescription)")
             fatalError("\(#file) \(#function) could not create image file name \(fileName)")
+        }
+    }
+    
+    func imageFileName(_ imageURL: URL?) -> String {
+        if let name = imageURL?.pathComponents.last {
+            return name
+        } else {
+            logger.info("\(#file) \(#function) error no file name for imageURL \(imageURL.debugDescription) ")
+            return ""
         }
     }
 }
