+++ PittsburghCityBridges/Views/BridgeDetailsView.swift
@@ -9,7 +9,7 @@ import SwiftUI
 import os
 
 struct BridgeDetailsView: View {
-    @ObservedObject var imageLoader: UIImageLoader
+    @ObservedObject var imageFileService: ImageFileService 
     @State private var bridgeImageOnly = false
     @State var bridgeImage = UIImage()
     @State private var dragOffset: CGSize = .zero
@@ -29,7 +29,7 @@ struct BridgeDetailsView: View {
     
     init(bridgeModel: BridgeModel) {
         self.bridgeModel = bridgeModel
-        imageLoader = UIImageLoader()
+        imageFileService = ImageFileService()
     }
     
     var dragGesture: some Gesture {
@@ -159,7 +159,7 @@ struct BridgeDetailsView: View {
                     UIScrollView.appearance().bounces = true
                     Task {
                         do {
-                            if let data = await imageLoader.getImageData(for: bridgeModel.imageURL),
+                            if let data = await imageFileService.getImageData(for: bridgeModel.imageURL),
                                let image =  UIImage(data: data) {
                                 bridgeImage = image
                             }
